Break up pip.index into smaller modules
pip/index.py had the most lines and 3 classes and a ton of imports.

I thought separating the classes might make things more understandable
and prevent merge conflicts.
